dashboard/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pyyaml>=6.0
psutil>=5.9.0
sse-starlette>=1.6.0
